requirements.txt
fitz==0.0.1.dev2
Pillow==8.2.0
Pillow==11.0.0
